Implement length conversion between yards and inches
This project is a simple C++ program that converts lengths between yards, feet, and inches, and total inches. It’s based on an example from Ivor Horton’s C++20: Beginner to Beyond, but adapted to be more user-friendly and modern by using format library that makes it easier for the user to put a variable inside a string.